# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EMARA API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginBody" }
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthUser" }
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      responses:
        "200":
          description: List clients
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Client" }
    post:
      operationId: createClient
      tags: [clients]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateClientBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Client" }
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Client" }
    patch:
      operationId: updateClient
      tags: [clients]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateClientBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Client" }
    delete:
      operationId: deleteClient
      tags: [clients]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204": { description: Deleted }
...
  /projects/{projectId}/approvals:
    get:
      operationId: listProjectApprovals
      tags: [approvals]
      parameters:
        - in: path
          name: projectId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Approvals
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ApprovalDetail" }
    post:
      operationId: requestApproval
      tags: [approvals]
      parameters:
        - in: path
          name: projectId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RequestApprovalBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApprovalDetail" }
...
  /approvals/{id}/respond:
    post:
      operationId: respondToApproval
      tags: [approvals]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RespondApprovalBody" }
      responses:
        "200":
          description: Responded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApprovalDetail" }
  /projects/{projectId}/comments:
    get:
      operationId: listProjectComments
      tags: [comments]
      parameters:
        - in: path
          name: projectId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CommentWithUser" }
    post:
      operationId: createProjectComment
      tags: [comments]
      parameters:
        - in: path
          name: projectId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCommentBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CommentWithUser" }
  /projects/{projectId}/boq:
    get:
      operationId: listProjectBoq
      tags: [boq]
      parameters:
        - in: path
          name: projectId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: BOQ items
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/BoqItem" }
    post:
      operationId: createBoqItem
      tags: [boq]
      parameters:
        - in: path
          name: projectId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateBoqItemBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BoqItem" }
...
  /boq-items/{id}:
    patch:
      operationId: updateBoqItem
      tags: [boq]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateBoqItemBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BoqItem" }
    delete:
      operationId: deleteBoqItem
      tags: [boq]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204": { description: Deleted }
...
components:
  schemas:
...
    Role:
      type: string
      enum: [admin, sales, designer, draftsman, qs, accountant, client]
    AuthUser:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        clientId: { type: ["integer", "null"] }
      required: [id, name, email, role, clientId]
    LoginBody:
      type: object
      properties:
        email: { type: string }
        password: { type: string }
      required: [email, password]
...
    Client:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        phone: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        source: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
      required: [id, name, phone, email, address, source, notes, createdAt]
    CreateClientBody:
      type: object
      properties:
        name: { type: string }
        phone: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        source: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
      required: [name]
    UpdateClientBody:
      type: object
      properties:
        name: { type: string }
        phone: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        source: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
...
    ApprovalStatus:
      type: string
      enum: [pending, approved, approved_with_comments, revision_required, rejected]
    ApprovalDetail:
      type: object
      properties:
        id: { type: integer }
        projectId: { type: integer }
        projectName: { type: string }
        clientName: { type: string }
        stageId: { type: integer }
        stageName: { type: string }
        requestedById: { type: integer }
        requestedByName: { type: string }
        approvedById: { type: ["integer", "null"] }
        status: { $ref: "#/components/schemas/ApprovalStatus" }
        clientComment: { type: ["string", "null"] }
        internalComment: { type: ["string", "null"] }
        requestedAt: { type: string, format: date-time }
        respondedAt: { type: ["string", "null"], format: date-time }
      required:
        - id
        - projectId
        - projectName
        - clientName
        - stageId
        - stageName
        - requestedById
        - requestedByName
        - approvedById
        - status
        - clientComment
        - internalComment
        - requestedAt
        - respondedAt
    RequestApprovalBody:
      type: object
      properties:
        stageId: { type: integer }
        internalComment: { type: ["string", "null"] }
      required: [stageId]
    RespondApprovalBody:
      type: object
      properties:
        status: { $ref: "#/components/schemas/ApprovalStatus" }
        clientComment: { type: ["string", "null"] }
      required: [status]
    CommentWithUser:
      type: object
      properties:
        id: { type: integer }
        projectId: { type: integer }
        stageId: { type: ["integer", "null"] }
        approvalId: { type: ["integer", "null"] }
        userId: { type: integer }
        userName: { type: string }
        userRole: { $ref: "#/components/schemas/Role" }
        commentText: { type: string }
        isInternal: { type: boolean }
        revisionNumber: { type: ["integer", "null"] }
        createdAt: { type: string, format: date-time }
      required:
        - id
        - projectId
        - stageId
        - approvalId
        - userId
        - userName
        - userRole
        - commentText
        - isInternal
        - revisionNumber
        - createdAt
    CreateCommentBody:
      type: object
      properties:
        stageId: { type: ["integer", "null"] }
        approvalId: { type: ["integer", "null"] }
        commentText: { type: string }
        isInternal: { type: boolean }
      required: [commentText, isInternal]
    BoqCategory:
      type: string
      enum:
        - demolition
        - masonry
        - plumbing
        - electrical
        - hvac
        - ceiling
        - flooring
        - painting
        - carpentry
        - aluminum_glass
        - furniture
        - kitchen
        - bathroom
        - accessories
        - other
    BoqUnit:
      type: string
      enum: [m2, linear_meter, piece, set, day, lump_sum]
    BoqItem:
      type: object
      properties:
        id: { type: integer }
        projectId: { type: integer }
        category: { $ref: "#/components/schemas/BoqCategory" }
        itemName: { type: string }
        description: { type: ["string", "null"] }
        unit: { $ref: "#/components/schemas/BoqUnit" }
        quantity: { type: number }
        unitPrice: { type: number }
        totalPrice: { type: number }
        supplierName: { type: ["string", "null"] }
        executionStage: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
      required:
        - id
        - projectId
        - category
        - itemName
        - description
        - unit
        - quantity
        - unitPrice
        - totalPrice
        - supplierName
        - executionStage
        - notes
        - createdAt
...
    CreateBoqItemBody:
      type: object
      properties:
        category: { $ref: "#/components/schemas/BoqCategory" }
        itemName: { type: string }
        description: { type: ["string", "null"] }
        unit: { $ref: "#/components/schemas/BoqUnit" }
        quantity: { type: number }
        unitPrice: { type: number }
        supplierName: { type: ["string", "null"] }
        executionStage: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
      required: [category, itemName, unit, quantity, unitPrice]
    UpdateBoqItemBody:
      type: object
      properties:
        category: { $ref: "#/components/schemas/BoqCategory" }
        itemName: { type: string }
        description: { type: ["string", "null"] }
        unit: { $ref: "#/components/schemas/BoqUnit" }
        quantity: { type: number }
        unitPrice: { type: number }
        supplierName: { type: ["string", "null"] }
        executionStage: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
...