# lib/api-zod/src/generated/types/boqItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EMARA API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BoqCategory } from "./boqCategory";
import type { BoqUnit } from "./boqUnit";

export interface BoqItem {
  id: number;
  projectId: number;
  category: BoqCategory;
  itemName: string;
  /** @nullable */
  description: string | null;
  unit: BoqUnit;
  quantity: number;
  unitPrice: number;
  totalPrice: number;
  /** @nullable */
  supplierName: string | null;
  /** @nullable */
  executionStage: string | null;
  /** @nullable */
  notes: string | null;
  createdAt: Date;
}